refactor(api-v4): clean up resources + create apis resource
