+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v4/gravitee-apim-rest-api-management-v4-rest/src/main/java/io/gravitee/rest/api/management/v4/rest/resource/installation/OrganizationResource.java
@@ -16,16 +16,10 @@
 package io.gravitee.rest.api.management.v4.rest.resource.installation;
 
 import io.gravitee.common.http.MediaType;
-import io.gravitee.rest.api.management.v4.rest.mapper.EnvironmentMapper;
 import io.gravitee.rest.api.management.v4.rest.mapper.OrganizationMapper;
-import io.gravitee.rest.api.management.v4.rest.model.Environment;
 import io.gravitee.rest.api.management.v4.rest.model.Organization;
 import io.gravitee.rest.api.management.v4.rest.resource.AbstractResource;
-import io.gravitee.rest.api.model.EnvironmentEntity;
-import io.gravitee.rest.api.model.OrganizationEntity;
-import io.gravitee.rest.api.service.EnvironmentService;
 import io.gravitee.rest.api.service.OrganizationService;
-import java.util.Collection;
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -48,20 +42,10 @@ public class OrganizationResource extends AbstractResource {
     @Inject
     private OrganizationService organizationService;
 
-    @Inject
-    private EnvironmentService environmentService;
-
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public Organization get(@PathParam("orgId") String orgId) {
         return OrganizationMapper.INSTANCE.convert(organizationService.findById(orgId));
-    }
-
-    @Path("/environments")
-    @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public Collection<Environment> getEnvironments(@PathParam("orgId") String orgId) {
-        return EnvironmentMapper.INSTANCE.convertCollection(this.environmentService.findByUser(orgId, getAuthenticatedUserOrNull()));
     }
 }
